refactor: also separate tests
